Prepare for .NET 10
Cherry-pick refactoring from #264 to support multi-targeting for `net10.0`.
