+++ .github/workflows/build.yml
@@ -152,6 +152,7 @@ jobs:
     if: |
       github.event.repository.fork == false &&
       (github.ref_name == github.event.repository.default_branch ||
+       github.head_ref == 'dotnet-vnext' ||
        startsWith(github.ref, 'refs/tags/v'))
 
     environment:
