+++ src/OpenApi.Extensions/OpenApiExampleAttribute`2.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) Martin Costello, 2024. All rights reserved.
 // Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
 
-using System.Text.Json.Serialization;
-using Microsoft.OpenApi.Any;
-
 namespace MartinCostello.OpenApi;
 
 #pragma warning disable CA1813
@@ -14,7 +11,7 @@ namespace MartinCostello.OpenApi;
 /// <typeparam name="TSchema">The type of the schema.</typeparam>
 /// <typeparam name="TProvider">The type of the example provider.</typeparam>
 [AttributeUsage(ExampleTargets.ValidTargets, AllowMultiple = true, Inherited = true)]
-public class OpenApiExampleAttribute<TSchema, TProvider> : Attribute, IOpenApiExampleMetadata
+public partial class OpenApiExampleAttribute<TSchema, TProvider> : Attribute, IOpenApiExampleMetadata
     where TProvider : IExampleProvider<TSchema>
 {
     /// <inheritdoc/>
@@ -27,8 +24,4 @@ public class OpenApiExampleAttribute<TSchema, TProvider> : Attribute, IOpenApiEx
     /// A <typeparamref name="TSchema"/> that should be used as the example.
     /// </returns>
     public virtual TSchema GenerateExample() => TProvider.GenerateExample();
-
-    /// <inheritdoc/>
-    IOpenApiAny IOpenApiExampleMetadata.GenerateExample(JsonSerializerContext context)
-        => ExampleFormatter.AsJson(GenerateExample(), context);
 }
