+++ src/OpenApi.Extensions/PublicAPI/net9.0/PublicAPI.Unshipped.txt
@@ -1,0 +1,1 @@
+#nullable enable
