+++ tests/OpenApi.Extensions.Tests/ExampleFormatterTests.Can_Serialize_Boolean_value=False.DotNet9_0.verified.txt
