+++ tests/OpenApi.Extensions.Tests/ExampleFormatterTests.Can_Serialize_Boolean_value=True.DotNet9_0.verified.txt
