+++ tests/OpenApi.Extensions.Tests/ExampleFormatterTests.Can_Serialize_Complex_Object.DotNet9_0.verified.txt
