+++ tests/OpenApi.Extensions.Tests/ExampleFormatterTests.Can_Serialize_Integer.DotNet9_0.verified.txt
