+++ tests/OpenApi.Extensions.Tests/ExampleFormatterTests.Can_Serialize_Long.DotNet9_0.verified.txt
