+++ tests/OpenApi.Extensions.Tests/ExampleFormatterTests.Can_Serialize_Null_String.DotNet9_0.verified.txt
