+++ tests/OpenApi.Extensions.Tests/IntegrationTests.Schema_Is_Correct_For_App.DotNet9_0.verified.txt
