+++ tests/OpenApi.Extensions.Tests/IntegrationTests.Schema_Is_Correct_For_App_As_Yaml.DotNet9_0.verified.txt
