+++ tests/OpenApi.Extensions.Tests/IntegrationTests.Schema_Is_Correct_For_Classes.DotNet9_0.verified.txt
