+++ tests/OpenApi.Extensions.Tests/IntegrationTests.Schema_Is_Correct_For_Example_Attribute_Hierarchy.DotNet9_0.verified.txt
