+++ tests/OpenApi.Extensions.Tests/IntegrationTests.Schema_Is_Correct_For_Not_Json.DotNet9_0.verified.txt
