+++ tests/OpenApi.Extensions.Tests/IntegrationTests.Schema_Is_Correct_For_Sample.DotNet9_0.verified.txt
